Test: move lax.full sharding tests to multi_device_test
